videos-49ivnlvufswn/videos/1-environment-starter-repo-quickstart/requirements.txt
faiss-cpu
sentence-transformers
tiktoken
python-dotenv
